refactor: adjust method step
