+++ src/app/assets/svg/index.ts
@@ -107,11 +107,17 @@ import ConfirmYourPassphrase from "./confirm-your-passphrase.svg?react";
 import PersonLight from "./person-light.svg?react";
 import PersonDark from "./person-dark.svg?react";
 import CircleMinus from "./circle-minus.svg?react";
+import LedgerImportMethod from "./ledger-import-method.svg?react";
+import MnemonicImportMethod from "./mnemonic-import-method.svg?react";
+import SecretImportMethod from "./secret-import-method.svg?react";
+import AddressImportMethod from "./address-import-method.svg?react";
+
 
 export const SvgCollection: Record<string, FC<SVGProps<SVGSVGElement>>> = {
 	...ArrowIcons,
 	...Currencies,
 	...TransactionIcons,
+	AddressImportMethod,
 	Bell,
 	Calendar,
 	Categories,
@@ -174,6 +180,7 @@ export const SvgCollection: Record<string, FC<SVGProps<SVGSVGElement>>> = {
 	ImportWalletLight,
 	Ledger,
 	LedgerAlt,
+	LedgerImportMethod,
 	List,
 	LoaderLogo,
 	Lock,
@@ -183,6 +190,7 @@ export const SvgCollection: Record<string, FC<SVGProps<SVGSVGElement>>> = {
 	Mainnet,
 	Menu,
 	MenuOpen,
+	MnemonicImportMethod,
 	MoneyCoinSwap,
 	NoteCheck,
 	Pencil,
@@ -192,6 +200,7 @@ export const SvgCollection: Record<string, FC<SVGProps<SVGSVGElement>>> = {
 	Plus,
 	QRCode,
 	QuestionMarkSmall,
+	SecretImportMethod,
 	SelectNetworkDark,
 	SelectNetworkLight,
 	Server,
