+++ src/domains/portfolio/components/ImportWallet/ImportWallet.WIF.test.tsx
@@ -1,13 +1,12 @@
-import { Contracts } from "../../../../../../platform-sdk/packages/profiles/source/helpers";
 import userEvent from "@testing-library/user-event";
 import { createHashHistory } from "history";
 import React from "react";
 import { FormProvider, useForm } from "react-hook-form";
 import { Route } from "react-router-dom";
 
-import { MethodStep } from "./MethodStep";
-import { EnvironmentProvider } from "../../../../app/contexts";
-import { OptionsValue } from "../../../wallet/hooks/use-import-options";
+import { ImportDetailStep } from "./ImportDetailStep";
+import { EnvironmentProvider } from "@/app/contexts";
+import { OptionsValue } from "@/domains/wallet/hooks/use-import-options";
 import {
 	env,
 	getDefaultProfileId,
@@ -16,8 +15,9 @@ import {
 	waitFor,
 	mockProfileWithPublicAndTestNetworks,
 	act,
-} from "../../../../utils/testing-library";
-import * as usePortfolio from "../../hooks/use-portfolio";
+} from "@/utils/testing-library";
+import * as usePortfolio from "@/domains/portfolio/hooks/use-portfolio";
+import { Contracts } from "@ardenthq/sdk-profiles";
 
 let profile: Contracts.IProfile;
 const fixtureProfileId = getDefaultProfileId();
@@ -83,7 +83,7 @@ describe("ImportWallet WIF", () => {
 		return (
 			<EnvironmentProvider env={env}>
 				<FormProvider {...form}>
-					<MethodStep profile={profile} network={network} />
+					<ImportDetailStep profile={profile} network={network} />
 				</FormProvider>
 			</EnvironmentProvider>
 		);
